+++ lending/RELEASES.md
@@ -509,3 +509,13 @@ Based on:
 - [typescript v0.5.5] lending
 ### Releases
 - [NPM v0.5.5] https://www.npmjs.com/package/ryan-lending/v/0.5.5 - lending
+
+## 2024-11-06 00:57:46
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.431.0 (2.451.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.6.0] lending
+### Releases
+- [NPM v0.6.0] https://www.npmjs.com/package/ryan-lending/v/0.6.0 - lending